Added Set, Match, and start_tournament()
Completed the rest of the model needed for storing information re-
lating to the tournaments and the progression of players.

The start_tournament() method in class tournament is useful for
creating the brackets and getting the database in a ready state for
the rest of the logic needed to run.